refactor : 텃밭 분양 상세보기 채팅방 아이디 필드 추가
